Set bounds for dependencies
This is a major bump for almost everything, as we can’t be certain we’re not
excluding something that used to work. However, `yaya-containers` escapes with a
minor bump, because it already had bounds and only excludes a version of
`containers` that could not have compiled. Also, constricting the bounds on test
suites doesn’t trigger a major bump.