Fold the test packages into existing lib packages
`yaya-test` was separate because the dependency on `yaya-hedgehog` prevented it
from being included in `yaya`. However, a separate test-suite package isn’t
particularly useful. This folds it into `yaya-hedgehog`, as the closest option.

Similarly, `yaya-unsafe-test` is folded into `yaya-unsafe`, which had no reason
to be separate in the first place.