Simplify `binop` and `filter`
By making them operations on the pattern functor instead of the fixed-point, we
eliminate some noise.

The change to `binop` is a clear win, as it shows that `binop` never cares about
the recursive structure. For `filter`, the recursive structure is unavoidable,
but the new version is simpler for most cases.