Restrict (co)recursive instances
This uses a `Strict` type family to tag types and type constructors with their
strictness, allowing instances to be constrained more precisely.

This mitigates the issue that using a non-strict functor with, say, `Mu` will
result in potentially infinite structures, since any functor used with `Mu` must
be `Strict` in order to be folded.